[TASK] Remove offset for fetching posts of a topic
The variable "posts" should contain all postings of a topic.
Currently it starts from the second posting (offset: 1) because
typo3.net wanted to have a special handing for the first posting.

If you need a to remove the first posting from the pagination
because you handle it separately, consider using something like
the slice-viewhelper from EXT:vhs.
https://github.com/FluidTYPO3/vhs/tree/development/Classes/ViewHelpers/Iterator