[BUGFIX] Use forum-object when adding/removing a topic
When adding a new topic the topic-counter in the forum was not
increased. Let the existing functions addTopic/removeTopic in the
forum-object take care of needed actions.

In topicFactory add the topic to a forum first, so that addPost()
can afterwards update the forum properly (set lastPost etc.).

When adding a new topic forum needs to take into account that a
lastPost might not yet exist (empty topic).