[!!!] Rename topicConformDeleteAction to deleteTopicAction
Rename this action from ModerationController as it does not contain a
confirm component and additionally contained a typo.

Important: As this patch changes the "switchableControllerActions" for
the plugin it is necessary to edit the plugin containing the "Forum"
action once, re-select "Forum" and save it in order to be able to delete
topics.