Update counters of forums in Counter scheduler task
When updating the counters in the Counter scheduler task no updates to
the topic- and post counters of forums were made. Managing records in
the TYPO3 backend or using the moderation feature may lead to wrong
counters for forums which must at least be corrected by the counter
task like done for topics and users.