[FEATURE] Allow defining of classes for RootlineViewHelper items
This patch allows to define classes for the icons of forum and topic in
the rootline (omitting the icons completely if not set) and allows to
use breadcrumb instead of nav-pills (which is configured as default for
the bootstrap templates).

This patch also fixes the "reverse" attribute, which did not work yet.