[BUGFIX] Show "disable" option for fe_users in backend again
Commit d795a992 `[TASK] reports & mod function` in 07-2013 added a TCA
config overriding the `fe_users.disable` column type from `check` to
`passthrough` which hides the field in the backend making it impossible
to disable fe_users records there. This was probably not on purpose as
there are no negative side effects when editing the record in the
backend.

This patch removes the override in order to re-enable the functionality
to disable fe_users in the backend.