[BUGFIX] Remove character set definitions for cache tables
A separate definition setting the character set for the cache tables is
not needed in any current version of TYPO3 as UTF-8 always is default.
In contrast it causes issues with databases configured to use `utf8mb4`.
In order to simply use the database default as all other tables do this
patch removes the character set definitions for the cache tables.