[!!!][BUGFIX] Show link to topic in topic subscription mail
This change makes post in postCreated and topic in topicCreated
available as persisted objects. When sending a notification
email for a newly created topic it now contains a link to the
topic. This was not possible before as the topic had not been
persisted yet when sending the mail.

This also circumvents issues with sending the mail leading to
the post or topic not being created.

This changes the behaviour for other subscribers as the post
and topic have already been persisted in the slots "topicCreated"
and "postCreated".